requirements.txt
affine==2.4.0
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.5
llvmlite==0.41.1
numba==0.58.1
numpy==1.26.2
pyparsing==3.1.1
PyYAML==6.0.1
rasterio==1.3.9
six==1.16.0
snuggs==1.4.7
tqdm==4.66.1
